Show result after 2s from when we start the stop spinning
Need to come back to this and not hardcode it. It's hard to compute the
time since the amount of time is based on the symbols we're generating
and then they interpolate to get to their positions, but there's enough
symbols that the time based on all those going on is way too long and
the symbols arrriving are probably hanging out a few decimal points off
rom their final destination for a while.

Anyway. Probably tired and need to sleep on it